refactor(labyrinth): move the reachability domain kinds into aether-labyrinth
Relocate the 15 aether.reach.* / aether.corridor.* domain kinds (plus
their 8 helper sub-structs) out of aether-kinds' control_plane module
into a new aether-labyrinth kinds module, re-exported at the crate root.
The aether.trajectory.* recorder kinds stay in aether-kinds. Every
#[kind(name = "...")] string is byte-identical through the move, so each
KindId (a hash of the name) is stable and no wire payload, schema, or
dispatch changes. The relocated NAME/ID + schema-shape assertions travel
with the kinds and still pass.

Re-point labyrinth's own imports (aether_kinds::X to crate::X for the
moved kinds, keeping aether_kinds::TrajectoryLog), the viewer's three
import sites plus its new aether-labyrinth path dep and doc comments, and
the aether-mesh surface_net doc link. Closes #1914.
